Add player literal from contexts and a has permission condition
A player literal will parse player texts which will then be derived from
context. There is also a new has permission condition, which checks if a
certain player/console has a permission. Also outer elements which
return a boolean, will now determine the execution of the program. When
a boolean element is false, it'll halt execution, acting as a sort of
'easy' if statement.